comments addresed

+++ pages/api/post/[id].ts
@@ -2,6 +2,11 @@ import { getSession } from "next-auth/client";
 import prisma from "../../../lib/prisma";
 
 export default async function handle(req, res) {
+  const session = await getSession({ req });
+  if (!session) {
+    res.status(401).json({ message: "Not authenticated" });
+    return;
+  }
   const postId = req.query.id;
   if (req.method === "DELETE") {
     const post = await prisma.post.delete({
